fix_nhwc_problem
