+++ diopi_test/python/conformance/diopi_runtime.py
@@ -180,6 +180,20 @@ def compute_nhwc_stride(size, itemsize=1, name=None):
     else:
         return compute_nhwc_stride_3d(size, itemsize)
 
+def set_nhwc(tensor_nchw, name_2d_3d):
+    ndim = tensor_nchw.ndim
+    if ndim == 3:
+        axis = (1, 2, 0)
+    elif ndim == 4 and name_2d_3d == '3d':
+        axis = (1, 2, 3, 0)
+    elif ndim == 4:
+        axis = (0, 2, 3, 1)
+    elif ndim == 5:
+        axis = (0, 2, 3, 4, 1)
+    nhwc_out = np.transpose(tensor_nchw, axis).copy()
+    nhwc_out.shape = tensor_nchw.shape
+    nhwc_out.strides = compute_nhwc_stride(tensor_nchw.shape, tensor_nchw.itemsize, name_2d_3d)
+    return nhwc_out
 
 def diopi_rt_init():
     init_library()
@@ -233,6 +247,9 @@ class Tensor(diopiTensor):
 
         if isinstance(size, (tuple, list)):
             size = Sizes(list(size))
+
+        if isinstance(stride, (tuple, list)):
+            stride = Sizes(list(stride))
 
         if data_ptr is None:
             diopiTensor.__init__(self, size, stride, dtype, device, context)
